Extract common writers logic into MqttPacketCodec
