+++ network/src/main/java/javasabr/mqtt/network/message/plain/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package javasabr.mqtt.network.message.plain;
+
+import org.jspecify.annotations.NullMarked;
